oh-tab-1xjo: refactor App.tsx into hooks
